[Wallet] Reorganize and add e2e tests
