+++ packages/mobile/e2e/src/utils/consts.js
@@ -1,0 +1,3 @@
+export const DEFAULT_RECIPIENT_ADDRESS = '0x22c8a9178841ba95a944afd1a1faae517d3f5daa'
+export const SAMPLE_BACKUP_KEY =
+  'general debate dial flock want basket local machine effort monitor stomach purity attend brand extend salon obscure soul open floor useful like cause exhaust'
